# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface UserResponse {
  id: number;
  username: string;
  coins: number;
}

export interface AuthResponse {
  user: UserResponse;
}

export type CreateWhisperRequestLifetime =
  (typeof CreateWhisperRequestLifetime)[keyof typeof CreateWhisperRequestLifetime];

export const CreateWhisperRequestLifetime = {
  "1h": "1h",
  "24h": "24h",
  "7d": "7d",
} as const;

export interface CreateWhisperRequest {
  /**
   * @minLength 1
   * @maxLength 500
   */
  content: string;
  lifetime: CreateWhisperRequestLifetime;
}

export interface ReactionsCount {
  fire: number;
  heart: number;
  wow: number;
}

export interface WhisperResponse {
  id: string;
  content: string;
  lifetime: string;
  expiresAt: string;
  createdAt: string;
  reactions: ReactionsCount;
  isOwn: boolean;
  replyCount: number;
}

export interface WhispersListResponse {
  whispers: WhisperResponse[];
  total: number;
  page: number;
  hasMore: boolean;
}

export type ReactionRequestType =
  (typeof ReactionRequestType)[keyof typeof ReactionRequestType];

export const ReactionRequestType = {
  fire: "fire",
  heart: "heart",
  wow: "wow",
} as const;

export interface ReactionRequest {
  type: ReactionRequestType;
}

export interface ReactionsResponse {
  reactions: ReactionsCount;
  /** Updated coins for the whisper owner (only if logged in and it's your whisper) */
  coins?: number;
}

export interface CreateReplyRequest {
  /**
   * @minLength 1
   * @maxLength 300
   */
  content: string;
}

export interface ReplyResponse {
  id: string;
  content: string;
  createdAt: string;
  isOwn: boolean;
}

export interface RepliesListResponse {
  replies: ReplyResponse[];
}

export interface StatsResponse {
  /** Total whispers that have died since the beginning */
  totalDied: number;
  /** Currently alive whispers */
  totalAlive: number;
  totalCreated: number;
}

export type ProfileResponseStats = {
  totalCreated: number;
  totalDied: number;
  totalReactionsReceived: number;
};

export interface ProfileResponse {
  user: UserResponse;
  whispers: WhisperResponse[];
  stats: ProfileResponseStats;
}

export interface ExtendRequest {
  /**
   * Hours to add (costs 5 coins per hour)
   * @minimum 1
   * @maximum 24
   */
  hours: number;
}

export interface MarketActionResponse {
  success: boolean;
  coinsSpent: number;
  coinsRemaining: number;
  newExpiresAt?: string;
  message: string;
}

export type GetWhispersParams = {
  page?: number;
  limit?: number;
};

export type GetTopWhispersParams = {
  limit?: number;
};