# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Shopot - Anonymous Secret Sharing API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: whispers
  - name: replies
  - name: stats
  - name: market
  - name: profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /whispers:
    get:
      operationId: getWhispers
      tags: [whispers]
      summary: Get active whispers feed
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhispersListResponse"
    post:
      operationId: createWhisper
      tags: [whispers]
      summary: Create a whisper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWhisperRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhisperResponse"

  /whispers/top:
    get:
      operationId: getTopWhispers
      tags: [whispers]
      summary: Get top whispers by reactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhispersListResponse"

  /whispers/{id}:
    get:
      operationId: getWhisper
      tags: [whispers]
      summary: Get a single whisper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhisperResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /whispers/{id}/react:
    post:
      operationId: reactToWhisper
      tags: [whispers]
      summary: React to a whisper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionsResponse"

  /whispers/{id}/replies:
    get:
      operationId: getReplies
      tags: [replies]
      summary: Get replies for a whisper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepliesListResponse"
    post:
      operationId: createReply
      tags: [replies]
      summary: Create a reply to a whisper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Whisper not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get global stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get own profile with whispers and stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /market/extend/{id}:
    post:
      operationId: extendWhisper
      tags: [market]
      summary: Spend coins to extend a whisper's life
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtendRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketActionResponse"
        "400":
          description: Not enough coins or whisper not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /market/kill/{id}:
    post:
      operationId: killWhisper
      tags: [market]
      summary: Spend coins to kill a whisper early
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketActionResponse"
        "400":
          description: Not enough coins or whisper not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    AuthRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 6
      required: [username, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
      required: [user]

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        coins:
          type: integer
      required: [id, username, coins]

    CreateWhisperRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
        lifetime:
          type: string
          enum: [1h, 24h, 7d]
      required: [content, lifetime]

    WhisperResponse:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        lifetime:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        reactions:
          $ref: "#/components/schemas/ReactionsCount"
        isOwn:
          type: boolean
        replyCount:
          type: integer
      required: [id, content, lifetime, expiresAt, createdAt, reactions, isOwn, replyCount]

    WhispersListResponse:
      type: object
      properties:
        whispers:
          type: array
          items:
            $ref: "#/components/schemas/WhisperResponse"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean
      required: [whispers, total, page, hasMore]

    ReactionsCount:
      type: object
      properties:
        fire:
          type: integer
        heart:
          type: integer
        wow:
          type: integer
      required: [fire, heart, wow]

    ReactionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [fire, heart, wow]
      required: [type]

    ReactionsResponse:
      type: object
      properties:
        reactions:
          $ref: "#/components/schemas/ReactionsCount"
        coins:
          type: integer
          description: Updated coins for the whisper owner (only if logged in and it's your whisper)
      required: [reactions]

    CreateReplyRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 300
      required: [content]

    ReplyResponse:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        isOwn:
          type: boolean
      required: [id, content, createdAt, isOwn]

    RepliesListResponse:
      type: object
      properties:
        replies:
          type: array
          items:
            $ref: "#/components/schemas/ReplyResponse"
      required: [replies]

    StatsResponse:
      type: object
      properties:
        totalDied:
          type: integer
          description: Total whispers that have died since the beginning
        totalAlive:
          type: integer
          description: Currently alive whispers
        totalCreated:
          type: integer
      required: [totalDied, totalAlive, totalCreated]

    ProfileResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        whispers:
          type: array
          items:
            $ref: "#/components/schemas/WhisperResponse"
        stats:
          type: object
          properties:
            totalCreated:
              type: integer
            totalDied:
              type: integer
            totalReactionsReceived:
              type: integer
          required: [totalCreated, totalDied, totalReactionsReceived]
      required: [user, whispers, stats]

    ExtendRequest:
      type: object
      properties:
        hours:
          type: integer
          minimum: 1
          maximum: 24
          description: Hours to add (costs 5 coins per hour)
      required: [hours]

    MarketActionResponse:
      type: object
      properties:
        success:
          type: boolean
        coinsSpent:
          type: integer
        coinsRemaining:
          type: integer
        newExpiresAt:
          type: string
          format: date-time
        message:
          type: string
      required: [success, coinsSpent, coinsRemaining, message]