# lib/api-zod/src/generated/types/createReplyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateReplyRequest {
  /**
   * @minLength 1
   * @maxLength 300
   */
  content: string;
}