# lib/api-zod/src/generated/types/createWhisperRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateWhisperRequestLifetime } from "./createWhisperRequestLifetime";

export interface CreateWhisperRequest {
  /**
   * @minLength 1
   * @maxLength 500
   */
  content: string;
  lifetime: CreateWhisperRequestLifetime;
}