# lib/api-zod/src/generated/types/extendRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */

export interface ExtendRequest {
  /**
   * Hours to add (costs 5 coins per hour)
   * @minimum 1
   * @maximum 24
   */
  hours: number;
}