# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */
import type { ProfileResponseStats } from "./profileResponseStats";
import type { UserResponse } from "./userResponse";
import type { WhisperResponse } from "./whisperResponse";

export interface ProfileResponse {
  user: UserResponse;
  whispers: WhisperResponse[];
  stats: ProfileResponseStats;
}