# lib/api-zod/src/generated/types/reactionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */
import type { ReactionRequestType } from "./reactionRequestType";

export interface ReactionRequest {
  type: ReactionRequestType;
}