# lib/api-zod/src/generated/types/reactionsCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */

export interface ReactionsCount {
  fire: number;
  heart: number;
  wow: number;
}