# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */

export interface StatsResponse {
  /** Total whispers that have died since the beginning */
  totalDied: number;
  /** Currently alive whispers */
  totalAlive: number;
  totalCreated: number;
}