# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */

export interface UserResponse {
  id: number;
  username: string;
  coins: number;
}