# lib/api-zod/src/generated/types/whisperResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopot - Anonymous Secret Sharing API
 * OpenAPI spec version: 0.2.0
 */
import type { ReactionsCount } from "./reactionsCount";

export interface WhisperResponse {
  id: string;
  content: string;
  lifetime: string;
  expiresAt: Date;
  createdAt: Date;
  reactions: ReactionsCount;
  isOwn: boolean;
  replyCount: number;
}